Use name and unit from FoodItem
Instead of the duplicated info on ShoppingListItem / Ingredient.

I duplicated to allow FoodItem to be updated and not affect Recipes
or ShoppingLists, but this means that unit could vary for Ingredients
with the same parent FoodItem, making calculating total quantities for
ShoppingLists more complicated.

To simplify, this commit updates all usages to refer to the parent
FoodItem fields, and a subsequent commit will remove the fields from
Ingredient and ShoppingListItem.