Soft delete FoodItems
Hard delete was not possible because Ingredients and ShoppingBasketItems
reference the table.

This commit adds a new column and UI components to allow soft deleting and
restoring.

It also removes the edit view since there is no need to edit a FoodItem's
name when you can delete it and create a new one instead.

Also removes the show view since all details are visible on the index view.