Use file from newly generated Rails 7 project
This commit replaces all of the boilerplate from a new Rails 6 app
with that from a Rails 7 app.

This seems to be the easiest way to make sure the app behaves exactly
as a newly generated Rails 7 project would.

I previously followed the official docs for upgrading, but this still
left webpack in use. It seemed like this was the only way to be fully
sure that nothing else would be left over and the app wouldn't be left
in a hybrid state.

After copying across the files I needed to run the following again to
deploy to Heroku:

bundle lock --add-platform x86_64-linux --add-platform ruby